pull out unwrap_rust_python result
