+++ kybra/compiler/kybra_generate/src/generators/unwrap_rust_python_result.rs
@@ -1,0 +1,17 @@
+pub fn generate() -> proc_macro2::TokenStream {
+    quote::quote! {
+        pub fn _kybra_unwrap_rust_python_result<T>(
+            rust_python_result: Result<T, PyRef<PyBaseException>>,
+            vm: &rustpython::vm::VirtualMachine
+        ) -> T {
+            match rust_python_result {
+                Ok(ok) => ok,
+                Err(err) => {
+                    let err_string: String = err.to_pyobject(vm).repr(vm).unwrap().to_string();
+
+                    panic!("{}", err_string);
+                },
+            }
+        }
+    }
+}
